+++ src/main/java/subway/domain/Sections.java
@@ -1,4 +1,4 @@
-package subway.line.domain;
+package subway.domain;
 
 import java.util.List;
 import java.util.Map;
