+++ src/main/java/subway/repository/StationDao.java
@@ -1,11 +1,10 @@
-package subway.station.dao;
+package subway.repository;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import subway.station.domain.Station;
-import subway.station.util.StationMapper;
+import subway.domain.Station;
 
 import javax.annotation.Resource;
 import java.sql.PreparedStatement;
