+++ src/main/java/subway/service/StationService.java
@@ -1,11 +1,11 @@
-package subway.station.service;
+package subway.service;
 
 import org.springframework.stereotype.Service;
-import subway.line.domain.Section;
-import subway.line.dao.SectionDao;
-import subway.line.domain.Sections;
-import subway.station.domain.Station;
-import subway.station.dao.StationDao;
+import subway.domain.Section;
+import subway.repository.SectionDao;
+import subway.domain.Sections;
+import subway.domain.Station;
+import subway.repository.StationDao;
 
 import javax.annotation.Resource;
 import java.util.ArrayList;
